+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.1.0]
 
+### Added
+
++ [[Backend-427]](https://www.notion.so/Talle-M-Envio-de-mail-una-vez-que-se-confirma-la-meet-7310a9bdde094a36ae8cea37a6f5ce5b?pvs=4)
++ Added Send of Mail to every interested when meet is confirmed
+
 ### Fixed
 
 + [[Backend-421]](https://www.notion.so/Talle-XS-Fix-DELETE-topics-id-7378756240dc41c9a39ca6c9f9a78786)